Fix bugs in entity lookup and records within records.
This addresses two issues. The first was an incorrect recheck when searching
for entities. The second has to do with how we store records within records,
which was not working properly since we were trying to deal with
the parent entity identifier on records but there is none since we
point down to records from the parent instead.