Changes to build a vdb.jar for use in external apks.
This changeset makes the project build a vdb.jar for use
in external APKs and fixes a few issues related to such usage.

In particular it converts schema registration to a service call,
moves the Authority constant to a new class, fixes cursors returned
by the AvroContentProviderProxy so they will work even if the subclass
is being accessed from another process, and fixes a bug with using
native URIs in queries.